tidy aggregator classes
