+++ lib/repor/aggregator/sum.rb
@@ -1,7 +1,7 @@
 module Repor
-  module Aggregators
-    class SumAggregator < BaseAggregator
-      def aggregation(groups)
+  module Aggregator
+    class Sum < Base
+      def aggregate(groups)
         relate(groups).select("SUM(#{expression}) AS #{sql_value_name}")
       end
     end
